// client/stocklinx app/src/redux/location/constant.ts
export enum locationConst {
  FETCH_LOCATIONS_REQUEST = "FETCH_LOCATIONS_REQUEST",
  FETCH_LOCATIONS_SUCCESS = "FETCH_LOCATIONS_SUCCESS",
  FETCH_LOCATIONS_FAILURE = "FETCH_LOCATIONS_FAILURE",
  FETCH_LOCATION_REQUEST = "FETCH_LOCATION_REQUEST",
  FETCH_LOCATION_SUCCESS = "FETCH_LOCATION_SUCCESS",
  FETCH_LOCATION_FAILURE = "FETCH_LOCATION_FAILURE",
  CREATE_LOCATION_REQUEST = "CREATE_LOCATION_REQUEST",
  CREATE_LOCATION_SUCCESS = "CREATE_LOCATION_SUCCESS",
  CREATE_LOCATION_FAILURE = "CREATE_LOCATION_FAILURE",
  UPDATE_LOCATION_REQUEST = "UPDATE_LOCATION_REQUEST",
  UPDATE_LOCATION_SUCCESS = "UPDATE_LOCATION_SUCCESS",
  UPDATE_LOCATION_FAILURE = "UPDATE_LOCATION_FAILURE",
  REMOVE_LOCATION_REQUEST = "REMOVE_LOCATION_REQUEST",
  REMOVE_LOCATION_SUCCESS = "REMOVE_LOCATION_SUCCESS",
  REMOVE_LOCATION_FAILURE = "REMOVE_LOCATION_FAILURE",
  SET_LOCATION = "SET_LOCATION",
  CLEAR_LOCATION = "CLEAR_LOCATION",
  SET_LOCATIONS = "SET_LOCATIONS",
  CLEAR_LOCATIONS = "CLEAR_LOCATIONS",
}